requirements.txt
altgraph==0.17.3
asttokens==2.2.1
auto-py-to-exe==2.34.0
autopep8==2.0.2
backcall==0.2.0
bottle==0.12.25
bottle-websocket==0.2.9
cffi==1.15.1
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
distlib==0.3.6
easygui==0.98.3
Eel==0.16.0
executing==1.2.0
filelock==3.8.2
fonttools==4.39.3
future==0.18.3
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.2
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
jupyter_client==8.1.0
jupyter_core==5.3.0
keyboard==0.13.5
kiwisolver==1.4.4
matplotlib==3.7.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.2
packaging==23.0
parso==0.8.3
pefile==2023.2.7
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
prompt-toolkit==3.0.38
psutil==5.9.4
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pyglet==2.0.3
Pygments==2.15.0
pyinstaller==5.10.1
pyinstaller-hooks-contrib==2023.2
pyparsing==3.0.9
python-dateutil==2.8.2
pywin32==306
pywin32-ctypes==0.2.0
pyzmq==25.0.2
six==1.16.0
stack-data==0.6.2
tornado==6.2
traitlets==5.9.0
virtualenv==20.17.1
wcwidth==0.2.6
whichcraft==0.6.1
zope.event==4.6
zope.interface==6.0
